Extração dos metodos estaticos do programa para a biblioteca
