--- BibliotecaNumerosNarcisista/Class1.cs
@@ -1,7 +1,0 @@
-﻿namespace BibliotecaNumerosNarcisista
-{
-    public class Class1
-    {
-
-    }
-}
